py/utils-validation/pydantic-notes/pyproject.toml
[project]
name = 'pydantic-notes'
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.2.3",
    "pydantic==1.10.6",
    "pyhumps>=3.8.0",
    "schema>=0.7.7",
    "typing_extensions==4.4.0",
]

[tool.uv]
dev-dependencies = ["ruff", "pylint"]

[tool.ruff.lint]
extend-select = ["I"]

[tool.pyright]
typeCheckingMode = "basic"
